+++ R/pumpCase.R
@@ -29,8 +29,8 @@ pumpCase.euclidean <- function(x, case = "address") {
 pumpCase.voronoi <- function(x, case = "address") {
   output <- x$statistic.data
   if (x$case.location == "address") {
-    lapply(output, function(x) cholera::fatalities.address$anchor[x == 1])
-  } else if (x$case.location == "nominal") {
+    lapply(output, function(x) cholera::ortho.proj$case[x == 1])
+  } else if (x$case.location == "fatality") {
     lapply(output, function(x) cholera::fatalities$case[x == 1])
   }
 }